+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - [10649](https://github.com/vegaprotocol/vega/issues/10649) - Ensure markets do not get stuck in liquidity auction after protocol upgrade.
 - [10641](https://github.com/vegaprotocol/vega/issues/10641) - Fix panic in amend during auction for isolated margin.
 - [10656](https://github.com/vegaprotocol/vega/issues/10656) - Fix funding rates bounds can be null for perpetuals.
+- [10601](https://github.com/vegaprotocol/vega/issues/10601) - Fix epoch by block height API is slow.
 
 ## 0.74.1
 
@@ -231,8 +232,8 @@
 - [10604](https://github.com/vegaprotocol/vega/issues/10604) - Register margin modes API subscriber.
 - [10595](https://github.com/vegaprotocol/vega/issues/10595) - Fix failed amends for isolated margin orders causing negative spread in console.
 - [10606](https://github.com/vegaprotocol/vega/issues/10606) - Party profiles `API` was not returning results.
-- [10625](https://github.com/vegaprotocol/vega/issues/10625) - Fix panic in update spot market. 
-- [9244](https://github.com/vegaprotocol/vega/issues/9244) - fix unit tests to use spot markets. 
+- [10625](https://github.com/vegaprotocol/vega/issues/10625) - Fix panic in update spot market.
+- [9244](https://github.com/vegaprotocol/vega/issues/9244) - fix unit tests to use spot markets.
 - [10625](https://github.com/vegaprotocol/vega/issues/10625) - Fix panic in update spot market.
 - [10649](https://github.com/vegaprotocol/vega/issues/10649) - Ensure markets do not get stuck in liquidity auction after protocol upgrade.